Improve reliability and check with Github Action
The main purpose of this change is to check the packaging and service works. To
do so, we rely on Github Actions to run several jobs to build the package,
install it, then check we can run the service, and run the smoke tests against
it.

Due to the limitations of Github Actions and its local version `act`, we can't
start the systemd service as it need a proper environment with priviledge that
we don't have in a such environment. Instead, we extract from the service
definition the starting command and use that one to start openhexa.

Along the trials and errors, we bumped in a few bugs that led to the rest of the
changes, improving the reliability of the package and service: the definition of
package dependencies, a post installation hook running the DB migration and
fixtures installation, the definition of healthcheck for each service, the use
of those healthcheck to start the systemd service, and a clean stop command.