Start and stop with one script, but also get status, and more
To ease the package OpenHexa and checks the requirements, we provide a first
script `./script/build.sh`.

To ease the usage of OpenHexa, but also to get in one place the way to do
several actions such as starting it, stoping it, getting its status, reporting
details about running services, updating the container images, or preparing
the database, we provide a second script `./script/openhexa.sh`. This script
can be used locally or when it's installed as a package. The definition of the
systemd service relies on it as the `postinst` package hook.

This also builds the smoke-tests container image so it can be used during the
Github Actions.