Depend on Debian Postgresql service
For the moment, all the third party services that OpenHexa relies on run into
a container. For local hosting, there isn't really a good reason to do so when
they might be managed by the local Linux distribution. This is the case of
PostgreSQL.

This change allows us to work with the PostgreSQL service managed by Debian. We
also experiment how we can remove some containers, as it adds some complexity,
and go to a classic system administration with less layers (that add their
part of complexity).

This also refactors and cleans up existing script code. A new script is also
added to ease the setup of OpenHexa (check the environment, purge an
installation, ...). Finally, the README has been updated to ease the onboarding.