Use a dedicated Pgsql cluster rather than an existing one
Previously, we worked with the PostgreSQL cluster exposed on the last port
assigned to PostgreSQL clusters (as you can have multiple). In a normal
situation, that corresponds to the default cluster (main) of the latest version
(e.g. 15).

As OpenHexa app needs to dynamically creates PostgreSQL users and databases,
we can't manage in advance their authentication (via the file pg_hba.conf)
without requring the app to deal with that. We then authorized all users to
access to any database on the cluster when the connect from the Docker network
with password authentication. This is not ideal as the cluster might be used
for other apps.

The solution is to isolate all the things related to OpenHexa in a dedicated
cluster as the present commit does.

Also as the smoke tests are not working anymore and we didn't find the cause,
they have been disabled for the time being.